+++ schematic/__main__.py
@@ -13,6 +13,7 @@ from schematic.schemas.commands import (
 from schematic.visualization.commands import (
     viz as viz_cli,
 )  # viz generation commands
+from schematic import __version__
 
 logger = logging.getLogger()
 click_log.basic_config(logger)
@@ -24,6 +25,7 @@ CONTEXT_SETTINGS = dict(help_option_names=["--help", "-h"])  # help options
 # invoke_without_command=True -> forces the application not to show aids before losing them with a --h
 @click.group(context_settings=CONTEXT_SETTINGS, invoke_without_command=True)
 @click_log.simple_verbosity_option(logger)
+@click.version_option(version=__version__, prog_name="schematic")
 def main():
     """
     Command line interface to the `schematic` backend services.
