+++ schematic/version.py
@@ -1,4 +1,4 @@
 """Sets the version of the package"""
 import importlib.metadata
 
-__version__ = importlib.metadata.version("schematic")
+__version__ = importlib.metadata.version("schematicpy")
